+++ src/main/java/com/feed_the_beast/ftbutilities/data/TeleportType.java
@@ -1,0 +1,11 @@
+package com.feed_the_beast.ftbutilities.data;
+
+public enum TeleportType
+{
+	HOME,
+	WARP,
+	BACK,
+	SPAWN,
+	TPA,
+	RTP,
+}
